pyproject.toml
[tool.poetry]
name = "srv-template"
version = "0.1.0"
description = "Service template"
authors = ["AnViSe <anvise@grodno.belpost.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.*"
dishka = "^1.4.2"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
structlog = "^24.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
